Added integer GET routines.
Added mechanism for reading the data bytes of a CAN frame sequentially.
This is now used to allow reading successive bytes as 8, 16, 24 and 32 bit
signed and unsigned integers.

Added routine to initialize a CAN frame descriptor.

All the routines that just operating on CAN frames without requiring any
library state are now in the new CAN_FRAME module.